Resolve dependency versions.
Fixes #9. This commit allows the resolver to be asynchronous, and changes the
default resolver to resolve the version of dependencies by loading the
package.json from unpkg. The package.json is also used to resolve the exact
version and file of a requested module such that equivalent requests (such as
for d3-array and d3-array@1) are coalesced.